two images -> one; launchModel; code style improved
