+++ SpaceXApp/Resources/Assets.xcassets/unknown.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "icons8-stellar.png",
+      "filename" : "Image.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
